implement version validation

+++ vns-cpp/version.hpp
@@ -2,10 +2,10 @@
 #define VERSION_HPP
 
 // vns main version
-constexpr int VERSION = 2;
+constexpr size_t VERSION = 2;
 // vns revision
-constexpr int REVISION = 3;
+constexpr size_t REVISION = 3;
 // vns patch
-constexpr int PATCH = 0;
+constexpr size_t PATCH = 0;
 
 #endif
